Edited

+++ app/core/bot/routers/user/callback.py
@@ -91,7 +91,7 @@ async def clbk_next(
             "push",
             value[0]
         )
-    except:
+    except BaseException:
         pass
 
     await log(callback)
@@ -144,7 +144,7 @@ async def clbk_success(
             "push",
             "100"
         )
-    except:
+    except BaseException:
         pass
 
     await log(callback)
@@ -198,7 +198,7 @@ async def clbk_back(
             reply_markup=keyboard_message,
             link_preview_options=link_opts
         )
-    except:
+    except BaseException:
         pass
 
     await log(callback)
@@ -280,11 +280,23 @@ async def clbk_cancel_confirm(
         action="msg_update",
         msg_id=callback.message.message_id
     )
-    if isinstance(msg_id, int) and msg_id != 0 and callback.message.bot:
+    msg_payment_id: User | bool | None | int = await manage_user(
+        tg_id=callback.from_user.id,
+        action="msg_payment_update",
+        msg_id=0
+    )
+    if (
+        isinstance(msg_id, int) and msg_id != 0 and callback.message.bot
+        and isinstance(msg_payment_id, int)
+    ):
         try:
             await callback.message.bot.delete_message(
                 callback.message.chat.id,
                 msg_id
+            )
+            await callback.message.bot.delete_message(
+                callback.message.chat.id,
+                msg_payment_id
             )
         except:
             pass
